refactor exports and files
